requirements.txt
langgraph
langchain
langchain-core
langchain-openai
langchain-anthropic
langsmith
pydantic
fastapi
uvicorn
python-dotenv
websockets
playwright
aiofiles>=23.0.0
aiosqlite>=0.19.0